refactor: move glossary fetching and building into separate files to make mockups in testing easier
